backend/requirements.txt
django~=5.0
djangorestframework~=3.14.0
django-environ
# psycopg2
psycopg2-binary==2.9.6
Pillow
django-apscheduler
django-cors-headers
face_recognition==1.3.0
numpy~=1.24.1
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow~=9.3.0
torch~=2.0.1
opencv-python~=4.7.0.72
torchvision~=0.15.2
tqdm~=4.64.1
tensorboardX~=2.6.1
easydict~=1.10
pip~=23.1.2
google~=3.0.0
cryptography~=39.0.0
lxml~=4.9.2
tornado~=6.2
setuptools~=65.5.0
pytz~=2022.7
cffi~=1.15.1
cvzone~=1.5.6
mediapipe~=0.10.2
dlib~=19.24.2
imutils~=0.5.4
scipy~=1.9.3
Werkzeug==2.3.7